refactor(runtime-tags): shell blocker reason codes; patch writers live in html/patch
